Reduced duplicate code, added a gh actions test step
