+++ .github/workflows/test.yml
@@ -1,4 +1,4 @@
-name: Test build
+name: Test
 
 on:
   pull_request:
@@ -26,8 +26,13 @@ jobs:
       - name: Set up JDK 20
         uses: actions/setup-java@v3
         with:
-          java-version: 20
+          java-version: '20'
           distribution: 'temurin'
+
+      - name: Test
+        uses: gradle/gradle-build-action@v2
+        with:
+          arguments: test
 
       - name: Build
         uses: gradle/gradle-build-action@v2
